+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * **trader:** added BNB top up feature and float, including buttons on PnL page.
 * **trader:** added Shut Down, Resume, and BVA buttons to Strategies page.
 * **trader:** added Reset buttons to Virtual Balances and PnL pages.
+* **trader:** decreased wallet buffer to 2%.
 
 ## [1.1.0] (2021-08-09)